Improvements: Extract security api module
