+++ extensions-api/src/main/scala/pl/touk/nussknacker/security/api/GlobalPermission.scala
@@ -1,0 +1,6 @@
+package pl.touk.nussknacker.security.api
+
+//TODO: Remove after admin tabs will removed
+object GlobalPermission {
+  type GlobalPermission = String
+}
