--- security/src/main/scala/pl/touk/nussknacker/ui/security/api/GlobalPermission.scala
@@ -1,7 +1,0 @@
-package pl.touk.nussknacker.ui.security.api
-
-object GlobalPermission {
-
-  type GlobalPermission = String
-
-}
