+++ ui/server/src/main/scala/pl/touk/nussknacker/ui/api/AuthorizeProcess.scala
@@ -4,7 +4,7 @@ import pl.touk.nussknacker.engine.api.process.ProcessId
 import pl.touk.nussknacker.ui.initialization.Initialization.nussknackerUser
 import pl.touk.nussknacker.ui.process.repository.FetchingProcessRepository
 import pl.touk.nussknacker.ui.security.api.LoggedUser
-import pl.touk.nussknacker.ui.security.api.Permission.Permission
+import pl.touk.nussknacker.security.api.Permission.Permission
 
 import scala.concurrent.{ExecutionContext, Future}
